requirements.txt
# Core AI/ML dependencies
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
diffusers>=0.24.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Modern CLI and utilities
typer[all]>=0.9.0
rich>=13.6.0
loguru>=0.7.0
tqdm>=4.66.0

# Video processing
ffmpeg-python>=0.2.0

# Configuration
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Optional: For advanced watermark detection
segment-anything>=1.0
groundingdino-py>=0.1.0
supervision>=0.16.0
